Fix typing

+++ skfem/assembly/basis/abstract_basis.py
@@ -32,7 +32,7 @@ class AbstractBasis:
     basis: List[Tuple[DiscreteField, ...]] = []
     X: ndarray
     W: ndarray
-    _sign: int = 1
+    _sign: float = 1.
 
     def __init__(self,
                  mesh: Mesh,
